Community

+++ game/client/clientshadowmgr.cpp
@@ -2314,7 +2314,7 @@ inline ShadowType_t CClientShadowMgr::GetActualShadowCastType( IClientRenderable
 class CShadowLeafEnum : public ISpatialLeafEnumerator
 {
 public:
-	bool EnumerateLeaf( int leaf, int context )
+	bool EnumerateLeaf( int leaf, intptr_t context )
 	{
 		m_LeafList.AddToTail( leaf );
 		return true;
