+++ game/shared/sceneimage.cpp
@@ -370,7 +370,7 @@ bool CSceneImage::CreateSceneImageFile( CUtlBuffer &targetBuffer, char const *pc
 
 	if ( !bQuiet )
 	{
-		Msg( "Scenes: String Table: %llu bytes\n", (uint64)(stringOffsets.Count() * sizeof( int )) );
+		Msg( "Scenes: String Table: %zu bytes\n", stringOffsets.Count() * sizeof( int ) );
 		Msg( "Scenes: String Pool: %d bytes\n", stringPool.TellMaxPut() );
 	}
